자바 폴더 파일명 규칙대로 변경
