+++ src/main/java/com/travelroulette/Dto/Country/CountryDto.java
@@ -1,4 +1,4 @@
-package com.travelroulette.dto.continent;
+package com.travelroulette.Dto.Country;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,8 +9,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class continentDto {
+public class CountryDto {
+    private String countryCode;
+    private String countryNameKor;
+    private String countryNameEng;
+    private String flagURL;
     private String continentNumber;
-    private String continentNameKor;
-    private String continentNameEng;
 }
