+++ src/main/java/com/travelroulette/Dto/Listofcountries/ListOfCountriesDto.java
@@ -1,4 +1,4 @@
-package com.travelroulette.dto.listofcountries;
+package com.travelroulette.Dto.Listofcountries;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class listOfCountriesDto {
+public class ListOfCountriesDto {
     private String countryCode;
     private String userId;
     private String checkContWishList;
