+++ src/main/java/com/travelroulette/Dto/User/userDto.java
@@ -1,4 +1,4 @@
-package com.travelroulette.dto.board;
+package com.travelroulette.Dto.User;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class boardDto {
-    private Integer boardNumber;
-    private String boardName;
+public class userDto {
+    private String userId;
+    private String pwd;
+    private String email;
+    private String gender;
 }
