+++ src/main/webapp/WEB-INF/web.xml
@@ -8,7 +8,7 @@
 <!--    이클립스 docker 톰캣에서 에러  주석건다면 여기서부터 -->
 		 <filter>
         <filter-name>EncodingFilter</filter-name>
-        <filter-class>com.travelroulette.filter.EncodingFilter</filter-class>
+        <filter-class>com.travelroulette.Filter.EncodingFilter</filter-class>
         <init-param>
             <param-name>encoding</param-name>
             <param-value>UTF-8</param-value>
